modified banyak sekali struktur UI
