+++ settings.html
@@ -1,0 +1,54 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>User Profile</title>
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css"
+          rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH"
+          crossorigin="anonymous">
+    <link rel="stylesheet" href="styles/styles.css">
+</head>
+<body>
+
+<button type="button" data-bs-toggle="modal" data-bs-target="#user-profile-box" data-bs-whatever="">Modal</button>
+
+<div class="user-profile-box modal" tabindex="-1" id="user-profile-box">
+    <div class="modal-dialog">
+        <div class="modal-content">
+            <div class="user-profile-box-header modal-header">
+                <img src="assets/header-icon-white.jpg"
+                     id="chat-box-logo-white">
+                <h2 class="modal-title" id="text-header"></h2>
+            </div>
+            <div class="modal-body">
+                <div class="user-profile-box-info">
+                    <img src="assets/default-profile-picture.jpg"
+                         id="photo-profile">
+                    <h5 id="username">Username</h5>
+                </div>
+
+                <div class="user-profile-box-change">
+                    <form id="edit-profile-form">
+                        <label for="new-photo">Change Photo</label><br>
+                        <input type="file" id="new-photo" name="new-photo"><br>
+                        <label for="new-username">Change Username</label><br>
+                        <input type="text" id="new-username" name="new-username"
+                               placeholder="Enter new username"><br>
+                        <button type="submit" id="save-button"><h4>Save</h4></button>
+                        <button type="button" id="back-button"><h4>Back to Chat</h4></button>
+                    </form>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+
+</body>
+
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
+        integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
+        crossorigin="anonymous"></script>
+<script src="scripts/js/settings.js"></script>
+
+</html>
