--- styles/style_profile.css
@@ -1,111 +1,0 @@
-body {
-    background-color: #9BDBFF;
-    font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-    margin: 0; /* Remove default margin */
-    padding: 0; /* Remove default padding */
-}
-.chat-container{
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: center;
-    height: 100vh; /* Set container height to viewport height */
-}
-#chat-icon{
-    width: 100px;
-    height: 100px;
-    border-radius: 10px;
-    cursor: pointer;
-    margin-right: 100px;
-}
-#user-profile-box{
-    width: 90%; /* Make chat box width 90% of the viewport */
-    max-width: 400px; /* Set maximum width */
-    background-color: #fff;
-    border-radius: 10px;
-    box-shadow: 0 0 20px rgba(0, 0, 0, 0.1);
-    z-index: 999;
-}
-.user-profile-box-header{
-    width: auto;
-    height: 40px;
-    display: flex;
-    flex-direction: row;
-    justify-content: left;
-    padding: 5%;
-    color: #fff;
-    background-color: #3DADFF;
-    border-bottom-left-radius: 0;
-    border-bottom-right-radius: 0;
-    border-top-left-radius: 10px;
-    border-top-right-radius: 10px;
-}
-#chat-box-logo-white{
-    width: auto;
-    height: 40px;
-    margin-right: 5%;
-}
-#text-header{
-    align-content: center;
-}
-.user-profile-box-info{
-    display: flex;
-    flex-direction: column;
-    margin-top: 3%;
-    align-items: center;
-}
-#photo-profile{
-    width: 80px;
-    height: 80px;
-    border-radius:50px;
-}
-.user-profile-box-change{
-    margin-left: 5%;
-    margin-top: 0%;
-}
-.user-profile-box-change label{
-    color: #413e3e;
-}
-.user-profile-box-change input{
-    border-color: #9d9d9d;
-}
-#new-username{
-    border-radius: 20px;
-}
-.user-profile-box-change button{
-    width: 140px;
-    height: 40px;
-    margin-top: 7%;
-    margin-left: 30%;
-    margin-bottom: 15px;
-    background-color: #3DADFF;
-    border: 0cm ;
-    border-radius: 10px;
-    color: #fff;
-    display: flex;
-    align-items: center;
-    cursor: pointer;
-    justify-content: center;
-}
-#save-button :hover{
-    color: #064069;
-}
-#back-button :hover{
-    color: #413e3e;
-}
-#back-button{
-    background-color: #9d9d9d;
-}
-@media screen and (max-width: 600px) {
-    .chat-container {
-        align-items: flex-start; /* Align items to the top */
-    }
-    #user-profile-box {
-        width: 90%; /* Make chat box width 90% of the viewport */
-        max-width: none; /* Remove max-width */
-        height: 90%; /* Adjust height for smaller screens */
-        overflow-y: auto; /* Add scroll if content exceeds height */
-        overflow-x: auto;
-        margin-left: 20px;
-    }
-}
