--- user_profile.html
@@ -1,43 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>User Profile</title>
-    <link rel="stylesheet" href="styles/style_profile.css">
-</head>
-<body>
-<div class="chat-container">
-    <div class="chat-icon">
-        <img src="assets/icon.jpg"
-             alt="Chat Icon"
-             id="chat-icon">
-    </div>
-
-    <div class="user-profile-box" id="user-profile-box">
-        <div class="user-profile-box-header">
-            <img src="assets/header-icon-white.jpg"
-                 id="chat-box-logo-white">
-            <h2 id="text-header"></h2>
-        </div>
-        <div class="user-profile-box-info">
-            <img src="assets/default-profile-picture.jpg"
-                 id="photo-profile">
-            <h5 id="username">Username</h5>
-        </div>
-        <div class="user-profile-box-change">
-            <form id="edit-profile-form">
-                <label for="new-photo">Change Photo</label><br>
-                <input type="file" id="new-photo" name="new-photo"><br>
-                <label for="new-username">Change Username</label><br>
-                <input type="text" id="new-username" name="new-username"
-                       placeholder="Enter new username"><br>
-                <button type="submit" id="save-button"><h4>Save</h4></button>
-                <button type="button" id="back-button"><h4>Back to Chat</h4></button>
-            </form>
-        </div>
-    </div>
-</div>
-</body>
-<script src="scripts/js/profile.js"></script>
-</html>
