+++ username.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Title</title>
+    <title>Chat-Box</title>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css"
           rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH"
           crossorigin="anonymous">
@@ -11,9 +11,24 @@
 </head>
 <body>
 
-<div id="main-container" class="container-fluid">
-    <iframe id="main-iframe" src="username.html"></iframe>
-    <button class="btn" id="fab-username"></button>
+<div class="chat-box" id="chat-box-username">
+    <div id="chat-box-header">
+        <img src="assets/header-icon-black.jpg"
+             id="chat-box-logo">
+        <h2>Chat-Box</h2>
+    </div>
+
+    <div id="chat-box-form">
+        <h5>Username</h5>
+        <form id="username-form">
+            <input id="username-input" type="text" placeholder="Enter username" name="username"><br>
+            <button class="btn" id="start-chat-button" type="submit">Start Chat</button>
+        </form>
+    </div>
+
+    <div id="chat-box-footer">
+        <h7><b>Made by Kelompok 1</b></h7>
+    </div>
 </div>
 
 </body>
@@ -21,6 +36,6 @@
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
         integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
         crossorigin="anonymous"></script>
-<script src="scripts/js/index.js"></script>
+<script src="scripts/js/username.js"></script>
 
 </html>
